+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Add API reference docs generation.
+- Add status fields to AstarteDefaultIngress.
 
 ## [1.0.1] - 2021-12-17
 ### Changed